Use 'template' keyword for getParameter<>() call
`ParameterSet` is not a templated class, but `getParameter<>()` method
is templated. If the method is called in a templated class via some
indirection (e.g., via array element):

    a[i].getParameter<int>("...")

`template` keyword is needed for Clang pre-3.6 (65d8b4c).

I believe this is wrong, but having `template` keyword should not harm
the code. It tell compiler that `getParameter<>()` name needs to be
looked up as a template.

See PR22247: http://llvm.org/bugs/show_bug.cgi?id=22247

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>