Mark 'problem_t' and 'bpmpd_problem_t' public constants as needed
According to Clang some of constants in `problem_t` (`infinity` and
`maximize`) are never used in the code.

This is in C++ file thus technically not visible outside this
translation unit and safe to be removed, but they might contain
useful information for future. If deemed as not needed, we can remove
them later on.

NOTE: #pragma diagnostic does not/work correctly with GCC C++
Preprocessor.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>